tools/requirements.txt
# iRacing Relay Server Python Dependencies
# Python 3.11+ required

# iRacing SDK wrapper
pyirsdk>=1.3.0

# WebSocket server
websockets>=12.0
